% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth-helpers}
\alias{auth-helpers}
\alias{get_token}
\alias{set_token}
\alias{get_key}
\alias{set_key}
\title{Manage Zengine™ API key and PAT}
\usage{
get_key()
set_key()
get_token()
set_token()
}
\value{
\code{get_key()} and \code{get_token()} return a character vector with the Zengine™ API key and PAT stored in \code{.Renviron}, respectively. If these values are not stored in \code{.Renviron}, the functions return \code{NULL}.

\code{set_key()} and \code{set_token()} invisibly return \code{NULL}.
}
\description{
These functions retrieve and store your Zengine™ API key and personal access token (PAT) in \code{.Renviron}, with help from \code{usethis::edit_r_environ()}.
}
