% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{get_zen}
\alias{get_zen}
\title{Get Zengine API data}
\usage{
get_zen(resource, id = NULL, record = NULL, limit = NULL)
}
\arguments{
\item{resource}{A character vector specifying a Zengine API resource. Common choices include \code{"forms"} and \code{"workspaces"}. A full list of resources is available \href{https://zenginehq.github.io/developers/rest-api/resources/}{here}. Defaults to \code{"forms"}.}

\item{id}{A character vector specifying a resource ID. Defaults to \code{NULL}.}

\item{record}{A character or numeric vector specifying a record linked to a resource ID. The \code{id} argument must be provided if \code{record} is used. Defaults to \code{NULL}.}

\item{limit}{An integer vector of the number of results returned on a \code{get_zen()} query. Defaults to \code{NULL}.}
}
\value{
A list of class \code{zengine}.
}
\description{
\code{get_zen()} gets data from the Zengine API.
}
