% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_wide.R
\name{unnest_wide}
\alias{unnest_wide}
\title{Unnest all list-cols in a data frame into columns}
\usage{
unnest_wide(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A tibble.
}
\description{
\code{unnest_wide()} unnests all list-cols in a data frame into atomic columns for each unique element in a list-col. Unnested columns are appended to the end of the dataset, and are distinguished by concatenating an underscore and a sequential number to the original list-col name.
}
