% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-usethis.R
\name{usethis-helper}
\alias{usethis-helper}
\alias{edit_r_environ}
\title{Edit .Renviron}
\usage{
edit_r_environ()
}
\description{
See \code{usethis::\link[usethis]{edit_r_environ}} for details.
}
\details{
This function opens \code{.Renviron} where you can store your Zengine™ API
key and personal access token (PAT). It is called by \code{set_key()} and \code{set_token()}.
}
\keyword{internal}
