% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomp.R
\name{whomp}
\alias{whomp}
\title{Whomp Zengine API lists into tibbles}
\usage{
whomp(.list, ...)
}
\arguments{
\item{.list}{A list of class \code{zengine} created by \code{get_zen()}.}

\item{...}{A comma separated list of unquoted variable names. \href{https://dplyr.tidyverse.org/reference/select_helpers.html}{Select helpers} from dplyr are supported.}
}
\value{
A tibble.
}
\description{
\code{whomp()} transforms lists obtained from the Zengine API into analysis-ready tibbles.
}
