% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizehiver-package.R
\docType{package}
\name{wizehiver-package}
\alias{wizehiver}
\alias{wizehiver-package}
\title{wizehiver: wizehiver: An R Wrapper for WizeHive's Zengine API}
\description{
An R wrapper for WizeHive's Zengine API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ivelasq.github.io/wizehiver}
  \item \url{https://github.com/ivelasq/wizehiver}
  \item Report bugs at \url{https://github.com/ivelasq/wizehiver/issues}
}

}
\author{
\strong{Maintainer}: Isabella Velásquez \email{ivelasq@gmail.com}

Authors:
\itemize{
  \item Gustavo Velásquez \email{gvelasq2github@gmail.com}
}

}
\keyword{internal}
